API configuration ventas and detalle ventas

+++ src/main/java/co/edu/unbosque/ciclo3back/dao/detalle_ventasDAO.java
@@ -3,6 +3,6 @@ package co.edu.unbosque.ciclo3back.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import co.edu.unbosque.ciclo3back.model.detalle_ventas;
 
-public interface detalle_ventasDAO extends JpaRepository<detalle_ventas, Integer> {
+public interface detalle_ventasDAO extends JpaRepository<detalle_ventas, Long> {
 
 }
